+++ src/Krystal/Db/Sql/QueryBuilderInterface.php
@@ -38,6 +38,17 @@ interface QueryBuilderInterface
     public function insert($table, array $data, $ignore = false);
 
     /**
+     * Builds and appends INSERT statement without explicit column names
+     * 
+     * @param string $table
+     * @param array $values Values to be inseted in columns
+     * @param boolean $ignore Whether to ignore when PK collisions occur
+     * @throws \LogicException if $values array is empty
+     * @return \Krystal\Db\Sql\QueryBuilder
+     */
+    public function insertShort($table, array $values, $ignore = false);
+
+    /**
      * Generate INSERT query for many records
      * 
      * @param string $table